chore: clean up, polish
