+++ app/components/SiteHeader.vue
@@ -40,7 +40,7 @@ header {
 }
 
 p.site-author {
-  @apply text-5xl font-display font-semibold;
+  @apply text-5xl font-display font-semibold md:whitespace-nowrap;
 }
 
 .avatar-wrapper {
